requirements.txt
NBT
anvil-parser
argparse